+++ src/entities/rectangle.rs
@@ -199,7 +199,7 @@ impl Entity for Rectangle {
         //not implemented
     }
 
-    fn render(&mut self, camera: &Camera, page: u32) {
+    fn render(&mut self, camera: &Camera, page: u16) {
 
         renderer::draw_rect(
             &self.model_rotated_points,
